feat: add WalletHeader separate app general Header component from specific pages Header component
